Get sqlx installed/running, rename databases modules to hosted_db modules ahead of adding DB persistence
