+++ src/lib.rs
@@ -3,5 +3,5 @@
 //! `stacks` is a service that proxies embedded databases
 //! like SQLite and DuckDB.
 
-pub mod databases;
+pub mod hosted_db;
 pub mod http;
