+++ src/main.rs
@@ -1,5 +1,5 @@
 use clap::{arg, command, value_parser, Command};
-use stacks::databases::{run_query, DBType};
+use stacks::hosted_db::{run_query, DBType};
 use stacks::http::run_server;
 use std::path::PathBuf;
